+++ src/main/java/no/ntnu/idatt2106/repository/RentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.RentDAO;
 
-public interface RentRepository extends JpaRepository<RentDAO, Long> {
+public interface RentRepository extends JpaRepository<RentDAO, Integer> {
     
 }
